requirements.txt
annotated-types==0.6.0
anyio==4.2.0
backoff==2.2.1
boilerpy3==1.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
gradio==4.18.0
h11==0.14.0
haystack-ai==2.0.0b7
haystack-bm25==1.0.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.2.0
networkx==3.2.1
numpy==1.26.4
openai==1.12.0
pandas==2.2.0
posthog==3.4.1
pydantic==2.6.1
pydantic_core==2.16.2
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
